+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,9 +2,11 @@
 
 # HEAD (unreleased)
 
-- Fix(JsonEditor): Remove unnecessary `onModelChange` event emissions on init.
+## 34.1.0 (2021-01-20)
+
 - Enhancement: Add `blur` and `dateTimeSelected` outputs to `DateTimeComponent`.
 - Enhancement: Add `showIcons` input to `ToggleComponent` to display optional icons.
+- Fix(JsonEditor): Remove unnecessary `onModelChange` event emissions on init.
 - Fix(json-editor): allow 'null' type in schema.
 
 ## 34.0.1 (2020-12-17)